+++ apps/web/pages/apps/installation/[[...step]].tsx
@@ -474,7 +474,7 @@ export const getServerSideProps = async (context: GetServerSidePropsContext) => 
     const personalAccount = {
       id: user.id,
       name: user.name,
-      avatar: user.avatarUrl,
+      avatarUrl: user.avatarUrl,
       alreadyInstalled: appInstalls.some((install) => !Boolean(install.teamId) && install.userId === user.id),
     };
