+++ packages/features/auth/signup/utils/prefillAvatar.ts
@@ -1,9 +1,9 @@
 import type { Prisma } from "@prisma/client";
 import fetch from "node-fetch";
 
+import { uploadAvatar } from "@calcom/lib/server/avatar";
 import { resizeBase64Image } from "@calcom/lib/server/resizeBase64Image";
 import prisma from "@calcom/prisma";
-import { uploadAvatar } from "@calcom/trpc/server/routers/loggedInViewer/updateProfile.handler";
 
 interface IPrefillAvatar {
   email: string;
