+++ packages/lib/server/avatar.ts
@@ -2,6 +2,32 @@ import { v4 as uuidv4 } from "uuid";
 
 import { prisma } from "@calcom/prisma";
 
+export const uploadAvatar = async ({ userId, avatar: data }: { userId: number; avatar: string }) => {
+  const objectKey = uuidv4();
+
+  await prisma.avatar.upsert({
+    where: {
+      teamId_userId_isBanner: {
+        teamId: 0,
+        userId,
+        isBanner: false,
+      },
+    },
+    create: {
+      userId: userId,
+      data,
+      objectKey,
+      isBanner: false,
+    },
+    update: {
+      data,
+      objectKey,
+    },
+  });
+
+  return `/api/avatar/${objectKey}.png`;
+};
+
 export const uploadLogo = async ({
   teamId,
   logo: data,
