+++ packages/trpc/server/routers/loggedInViewer/updateProfile.handler.ts
@@ -2,7 +2,6 @@ import { Prisma } from "@prisma/client";
 // eslint-disable-next-line no-restricted-imports
 import { keyBy } from "lodash";
 import type { GetServerSidePropsContext, NextApiResponse } from "next";
-import { v4 as uuidv4 } from "uuid";
 
 import stripe from "@calcom/app-store/stripepayment/lib/server";
 import { getPremiumMonthlyPlanPriceId } from "@calcom/app-store/stripepayment/lib/utils";
@@ -12,6 +11,7 @@ import hasKeyInMetadata from "@calcom/lib/hasKeyInMetadata";
 import { HttpError } from "@calcom/lib/http-error";
 import logger from "@calcom/lib/logger";
 import { getTranslation } from "@calcom/lib/server";
+import { uploadAvatar } from "@calcom/lib/server/avatar";
 import { checkUsername } from "@calcom/lib/server/checkUsername";
 import { updateNewTeamMemberEventTypes } from "@calcom/lib/server/queries";
 import { resizeBase64Image } from "@calcom/lib/server/resizeBase64Image";
@@ -34,32 +34,6 @@ type UpdateProfileOptions = {
     res?: NextApiResponse | GetServerSidePropsContext["res"];
   };
   input: TUpdateProfileInputSchema;
-};
-
-export const uploadAvatar = async ({ userId, avatar: data }: { userId: number; avatar: string }) => {
-  const objectKey = uuidv4();
-
-  await prisma.avatar.upsert({
-    where: {
-      teamId_userId_isBanner: {
-        teamId: 0,
-        userId,
-        isBanner: false,
-      },
-    },
-    create: {
-      userId: userId,
-      data,
-      objectKey,
-      isBanner: false,
-    },
-    update: {
-      data,
-      objectKey,
-    },
-  });
-
-  return `/api/avatar/${objectKey}.png`;
 };
 
 export const updateProfileHandler = async ({ ctx, input }: UpdateProfileOptions) => {
