+++ packages/trpc/server/routers/viewer/workflows/filteredList.handler.tsx
@@ -1,7 +1,7 @@
 import type { WorkflowType } from "@calcom/ee/workflows/components/WorkflowListPage";
 import { hasFilter } from "@calcom/features/filters/lib/hasFilter";
 import type { PrismaClient } from "@calcom/prisma";
-import type { Prisma } from "@calcom/prisma/client";
+import { Prisma } from "@calcom/prisma/client";
 import { MembershipRole } from "@calcom/prisma/client";
 import type { TrpcSessionUser } from "@calcom/trpc/server/trpc";
 
@@ -15,34 +15,36 @@ type FilteredListOptions = {
   input: TFilteredListInputSchema;
 };
 
-const includedFields = {
-  activeOn: {
-    select: {
-      eventType: {
-        select: {
-          id: true,
-          title: true,
-          parentId: true,
-          _count: {
-            select: {
-              children: true,
+const { include: includedFields } = Prisma.validator<Prisma.WorkflowDefaultArgs>()({
+  include: {
+    activeOn: {
+      select: {
+        eventType: {
+          select: {
+            id: true,
+            title: true,
+            parentId: true,
+            _count: {
+              select: {
+                children: true,
+              },
             },
           },
         },
       },
     },
-  },
-  steps: true,
-  team: {
-    select: {
-      id: true,
-      slug: true,
-      name: true,
-      members: true,
-      logo: true,
+    steps: true,
+    team: {
+      select: {
+        id: true,
+        slug: true,
+        name: true,
+        members: true,
+        logoUrl: true,
+      },
     },
   },
-};
+});
 
 export const filteredListHandler = async ({ ctx, input }: FilteredListOptions) => {
   const { prisma, user } = ctx;
